requirements.txt
bcrypt==4.3.0
dotenv==0.9.9
fastapi==0.115.12
fastapi-mail==1.4.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyjwt==2.10.1
python-multipart==0.0.20
sqlalchemy==2.0.41
uvicorn==0.34.2